requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.1.0
langchain-core>=0.2.0
streamlit>=1.27.2
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
